Move AppCard to its own component
